Remove an unnecessary service creation service.
