+++ subprojects/composite-builds/src/main/java/org/gradle/composite/internal/BuildStateFactory.java
@@ -19,7 +19,6 @@ package org.gradle.composite.internal;
 import org.gradle.api.artifacts.component.BuildIdentifier;
 import org.gradle.api.internal.BuildDefinition;
 import org.gradle.initialization.BuildCancellationToken;
-import org.gradle.internal.build.BuildModelControllerServices;
 import org.gradle.internal.build.BuildLifecycleControllerFactory;
 import org.gradle.internal.build.BuildState;
 import org.gradle.internal.build.RootBuildState;
@@ -36,7 +35,6 @@ import org.gradle.util.Path;
 public class BuildStateFactory {
     private final BuildTreeController buildTreeController;
     private final BuildLifecycleControllerFactory buildLifecycleControllerFactory;
-    private final BuildModelControllerServices buildModelControllerServices;
     private final ListenerManager listenerManager;
     private final GradleUserHomeScopeServiceRegistry userHomeDirServiceRegistry;
     private final CrossBuildSessionState crossBuildSessionState;
@@ -44,14 +42,12 @@ public class BuildStateFactory {
 
     public BuildStateFactory(BuildTreeController buildTreeController,
                              BuildLifecycleControllerFactory buildLifecycleControllerFactory,
-                             BuildModelControllerServices buildModelControllerServices,
                              ListenerManager listenerManager,
                              GradleUserHomeScopeServiceRegistry userHomeDirServiceRegistry,
                              CrossBuildSessionState crossBuildSessionState,
                              BuildCancellationToken buildCancellationToken) {
         this.buildTreeController = buildTreeController;
         this.buildLifecycleControllerFactory = buildLifecycleControllerFactory;
-        this.buildModelControllerServices = buildModelControllerServices;
         this.listenerManager = listenerManager;
         this.userHomeDirServiceRegistry = userHomeDirServiceRegistry;
         this.crossBuildSessionState = crossBuildSessionState;
@@ -59,17 +55,17 @@ public class BuildStateFactory {
     }
 
     public RootBuildState createRootBuild(BuildDefinition buildDefinition) {
-        return new DefaultRootBuildState(buildDefinition, buildTreeController, buildLifecycleControllerFactory, buildModelControllerServices, listenerManager);
+        return new DefaultRootBuildState(buildDefinition, buildTreeController, buildLifecycleControllerFactory, listenerManager);
     }
 
     public StandAloneNestedBuild createNestedBuild(BuildIdentifier buildIdentifier, Path identityPath, BuildDefinition buildDefinition, BuildState owner) {
-        return new DefaultNestedBuild(buildIdentifier, identityPath, buildDefinition, owner, buildTreeController, buildLifecycleControllerFactory, buildModelControllerServices);
+        return new DefaultNestedBuild(buildIdentifier, identityPath, buildDefinition, owner, buildTreeController, buildLifecycleControllerFactory);
     }
 
     public RootOfNestedBuildTree createNestedTree(BuildDefinition buildDefinition,
                                                   BuildIdentifier buildIdentifier,
                                                   Path identityPath,
                                                   BuildState owner) {
-        return new RootOfNestedBuildTree(buildDefinition, buildIdentifier, identityPath, owner, buildModelControllerServices, userHomeDirServiceRegistry, crossBuildSessionState, buildCancellationToken);
+        return new RootOfNestedBuildTree(buildDefinition, buildIdentifier, identityPath, owner, userHomeDirServiceRegistry, crossBuildSessionState, buildCancellationToken);
     }
 }
