+++ subprojects/composite-builds/src/main/java/org/gradle/composite/internal/DefaultIncludedBuildFactory.java
@@ -19,7 +19,6 @@ package org.gradle.composite.internal;
 import org.gradle.api.InvalidUserDataException;
 import org.gradle.api.artifacts.component.BuildIdentifier;
 import org.gradle.api.internal.BuildDefinition;
-import org.gradle.internal.build.BuildModelControllerServices;
 import org.gradle.internal.build.BuildLifecycleControllerFactory;
 import org.gradle.internal.build.BuildState;
 import org.gradle.internal.build.IncludedBuildFactory;
@@ -36,18 +35,15 @@ public class DefaultIncludedBuildFactory implements IncludedBuildFactory {
     private final Instantiator instantiator;
     private final WorkerLeaseService workerLeaseService;
     private final BuildLifecycleControllerFactory buildLifecycleControllerFactory;
-    private final BuildModelControllerServices buildModelControllerServices;
 
     public DefaultIncludedBuildFactory(BuildTreeController buildTree,
                                        Instantiator instantiator,
                                        WorkerLeaseService workerLeaseService,
-                                       BuildLifecycleControllerFactory buildLifecycleControllerFactory,
-                                       BuildModelControllerServices buildModelControllerServices) {
+                                       BuildLifecycleControllerFactory buildLifecycleControllerFactory) {
         this.buildTree = buildTree;
         this.instantiator = instantiator;
         this.workerLeaseService = workerLeaseService;
         this.buildLifecycleControllerFactory = buildLifecycleControllerFactory;
-        this.buildModelControllerServices = buildModelControllerServices;
     }
 
     private void validateBuildDirectory(File dir) {
@@ -71,8 +67,7 @@ public class DefaultIncludedBuildFactory implements IncludedBuildFactory {
             owner,
             buildTree,
             workerLeaseService.getCurrentWorkerLease(),
-            buildLifecycleControllerFactory,
-            buildModelControllerServices
+            buildLifecycleControllerFactory
         );
     }
 }
