+++ subprojects/composite-builds/src/main/java/org/gradle/composite/internal/RootOfNestedBuildTree.java
@@ -33,7 +33,6 @@ import org.gradle.internal.InternalBuildAdapter;
 import org.gradle.internal.build.AbstractBuildState;
 import org.gradle.internal.build.BuildLifecycleController;
 import org.gradle.internal.build.BuildLifecycleControllerFactory;
-import org.gradle.internal.build.BuildModelControllerServices;
 import org.gradle.internal.build.BuildState;
 import org.gradle.internal.build.BuildStateRegistry;
 import org.gradle.internal.build.NestedRootBuild;
@@ -73,7 +72,6 @@ public class RootOfNestedBuildTree extends AbstractBuildState implements NestedR
                                  BuildIdentifier buildIdentifier,
                                  Path identityPath,
                                  BuildState owner,
-                                 BuildModelControllerServices buildModelControllerServices,
                                  GradleUserHomeScopeServiceRegistry userHomeDirServiceRegistry,
                                  CrossBuildSessionState crossBuildSessionState,
                                  BuildCancellationToken buildCancellationToken) {
@@ -90,7 +88,6 @@ public class RootOfNestedBuildTree extends AbstractBuildState implements NestedR
         // Create the controller using the services of the nested tree
         BuildLifecycleControllerFactory buildLifecycleControllerFactory = buildTree.getServices().get(BuildLifecycleControllerFactory.class);
         BuildScopeServices buildScopeServices = new BuildScopeServices(buildTree.getServices());
-        buildModelControllerServices.supplyBuildScopeServices(buildScopeServices);
         this.buildLifecycleController = buildLifecycleControllerFactory.newInstance(buildDefinition, this, owner.getMutableModel(), buildScopeServices);
     }
