+++ subprojects/configuration-cache/src/main/kotlin/org/gradle/configurationcache/ConfigurationCacheServices.kt
@@ -16,6 +16,11 @@
 
 package org.gradle.configurationcache
 
+import org.gradle.api.internal.GradleInternal
+import org.gradle.configuration.ProjectsPreparer
+import org.gradle.configurationcache.build.ConfigurationCacheIncludedBuildState
+import org.gradle.configurationcache.build.NoOpBuildModelController
+import org.gradle.configurationcache.extensions.get
 import org.gradle.configurationcache.fingerprint.ConfigurationCacheFingerprintController
 import org.gradle.configurationcache.initialization.ConfigurationCacheBuildEnablement
 import org.gradle.configurationcache.initialization.ConfigurationCacheStartParameter
@@ -23,6 +28,13 @@ import org.gradle.configurationcache.initialization.DefaultConfigurationCachePro
 import org.gradle.configurationcache.initialization.DefaultInjectedClasspathInstrumentationStrategy
 import org.gradle.configurationcache.problems.ConfigurationCacheProblems
 import org.gradle.configurationcache.serialization.beans.BeanConstructors
+import org.gradle.initialization.ConfigurationCache
+import org.gradle.initialization.ConfigurationCacheAwareBuildModelController
+import org.gradle.initialization.SettingsPreparer
+import org.gradle.initialization.TaskExecutionPreparer
+import org.gradle.initialization.VintageBuildModelController
+import org.gradle.internal.build.BuildModelController
+import org.gradle.internal.build.BuildState
 import org.gradle.internal.event.ListenerManager
 import org.gradle.internal.service.ServiceRegistration
 import org.gradle.internal.service.scopes.AbstractPluginServiceRegistry
@@ -52,7 +64,6 @@ class ConfigurationCacheServices : AbstractPluginServiceRegistry() {
             add(ConfigurationCacheReport::class.java)
             add(ConfigurationCacheProblems::class.java)
             add(ConfigurationCacheClassLoaderScopeRegistryListener::class.java)
-            add(DefaultBuildModelControllerServices::class.java)
             add(DefaultConfigurationCacheProblemsListener::class.java)
         }
     }
@@ -64,6 +75,7 @@ class ConfigurationCacheServices : AbstractPluginServiceRegistry() {
             add(SystemPropertyAccessListener::class.java)
             add(RelevantProjectsRegistry::class.java)
             add(ConfigurationCacheFingerprintController::class.java)
+            addProvider(BuildScopeServicesProvider())
         }
     }
 
@@ -73,6 +85,24 @@ class ConfigurationCacheServices : AbstractPluginServiceRegistry() {
             add(ConfigurationCacheHost::class.java)
             add(ConfigurationCacheIO::class.java)
             add(DefaultConfigurationCache::class.java)
+        }
+    }
+
+    class BuildScopeServicesProvider {
+        fun createBuildModelController(build: BuildState, gradle: GradleInternal, startParameter: ConfigurationCacheStartParameter): BuildModelController {
+            if (build is ConfigurationCacheIncludedBuildState) {
+                return NoOpBuildModelController(gradle)
+            }
+            val projectsPreparer: ProjectsPreparer = gradle.services.get()
+            val settingsPreparer: SettingsPreparer = gradle.services.get()
+            val taskExecutionPreparer: TaskExecutionPreparer = gradle.services.get()
+            val configurationCache: ConfigurationCache = gradle.services.get()
+            val vintageController = VintageBuildModelController(gradle, projectsPreparer, settingsPreparer, taskExecutionPreparer)
+            return if (startParameter.isEnabled) {
+                ConfigurationCacheAwareBuildModelController(gradle, vintageController, configurationCache)
+            } else {
+                vintageController
+            }
         }
     }
 }
